ai-financial-advisorX-main/requirements.txt
streamlit==1.35.0
google-generativeai==0.7.2
python-dotenv==1.0.1
plotly==5.22.0
pandas>=2.2.3
numpy>=2.1.0
matplotlib
seaborn